Fixing for node 4.1.1
Not claiming this is the best fix, but it gets it working in the newest version of Node.
All the unit tests pass just fine, but the output files when it's run via CLI are empty, despite attempts to write the correct content to them.
This seems to be an issue with q-io, and using the Node-native FS module to write the file, in combination with Q's promises, resolves the issue.